Move snackbar up a level to survive dismantling KnackSync
